DEVPROD-12579: add support for conditional s3 puts (#117)
This commit adds support for conditional s3 puts to pail Buckets. This
feature can be controlled by setting the "IfNotExists" S3Options flag.
If the user does not set this flag to true, there is no change to writes.
If the flag is set to true, calls to .Close() or .Flush() will fail if
the key the user is trying to write to already exists.

I have added two tests that ensure this feature works by both checking
that we receive an error on write, and that the content we see in the S3
bucket matches what was originally written afterwards.

In order to enable this feature in the AWS S3 SDK, I had to bump the
version of the SDK we're using. This version bump also required a go.mod
version bump, but only to 1.21 from 1.20. The changes to our dependnecies
are only to support this version bump.

There was also apparently a change in the SDK to make the Content-Type
not populate a default if you pass a non-nil pointer to the appropriate
Input struct. Because some of our existing tests checked explicitly for
this, I made a small change to both bucket implementations to preserve
that behavior.