DEVPROD-16093: add FastGetS3Bucket interface for faster downloads
This commit adds a new interface, FastGetS3Bucket, which is a pail.Bucket but with
one additional method, GetToWriter. GetToWriter allows the caller to pass in an
io.WriterAt which will have the contents of the object automatically copied to
it. This method was added to enable support for AWS' official s3Manager.Download
API. This API is significantly more efficient for large files than the
s3.Client.GetObject() API alone because it uses byte ranges to fetch parts of
the object in parallel. It is not straightforward to use the existing Get
API with s3Manager.Download because of the io.WriterAt requirement.

In tests of a 35GB file I observed download times go from 25 to 7 minutes.
Even tests of smaller files consistently show improved download times.

I chose to make a new interface because I did not want to
potentially break any users of the Bucket interface, and adding support to the
other Bucket implementations felt pretty tricky and not worthwhile. Because our
constructors are returning an interface instead of a struct, it was necessary
to add a new constructor which returned this new interface. I'm not really
in love with this approach, but it's the safest for backward compatibility
given the current API of Pail.

I have included a test of this new behavior within the existing test fixtures.