DEVPROD-16611: Adjust concurrency for pail GetToWriter (#128)
This commit adjusts pail's GetToWriter API to use a higher level of concurrency
in all situations.

During my testing across a good number of Evergreen distros,
it seemed like 10, which is the aws CLI's default, was the best minimum
number. Even distros with fewer cores than 10 benefitted a good amount
by having the concurrency turned up to 10. For example, on rhel9-small
I saw the time to download a 35 GB file decrease from 9 minutes to 6 minutes.
The time it took to download a 3 GB file was about a second and a half
faster. An 80 MB file downloads a few hundred milliseconds faster.

The biggest improvement from this change is seen when we can go above the concurrency limit of 10. I tested this mostly on xlarge distros, where
setting the concurrency to the number of CPUs available dropped a 35 GB
file download to just about one minute compared to 4 minutes with the
AWS cli under default settings. I tested going above the CPU count and
noticed no significant improvement, so I am guessing there are other
bottlenecks we're hitting at that point. Much smaller file sizes were
approximately the same download time as the AWS CLI's default.

Given the huge variety of distros in use and file sizes we download,
there is likely a better formula we could work out. I don't believe this
change is the best we could possibly do, but I think it's significantly
better than always using a concurrency level of 5. If we want to optimize
this even further I think we would need a bigger investigation.