DEVPROD-16582 Use s3manager.Upload for s3Bucket.Put (#129)
This commit adjusts an s3Bucket's .Put method to use the aws SDK's
UploadManager API. This API transparently uploads files to S3 concurrently.
In my testing, this commit noticably improves upload speeds for larger files.
For example, on rhel90-small a 1GB upload dropped from 24.5 seconds to 9.7
seconds. The current implementation was OOM killed on a 16GB upload because
we buffer the entire file contents in memory before uploading. The new
implementation finished the upload in 180 seconds while using only minimal
memory. So, in addition to greatly improving performance, we are also enabling
large puts to run on smaller instance sizes.

One notable difference in how I have chosen to implement things here is
by using the existing .Put methods instead of implementing a new method.
I did this because we didn't really need a new API to efficiently use the
UploadManager, and I wanted to take advantage of all of the existing tests
to gain confidence in the implementation.

I have used the existing GetToWriter concurrency algorithm for tuning this
parameter because in my testing it seemed to have the same performance
curve as GetToWriter did.

One thing you'll notice is that we are still buffering the entire compressed
file contents in memory if Compress: true is set. I could not come up with
a satisfactory way of writing to disk in a robust manner for Evergreen's
use here while still allowing this feature. Even though s3.put seems to
never use it, I wanted to keep things compatible to minimize changes in
Evergreen. In the future we could look at either removing this feature or
changing other aspects of how it works to keep s3.put's uploads consistently
streamed from disk.