DEVPROD-17094: use crc32 checksum for s3 puts
This commit explicitly sets CRC32 as the checksum algorithm to use for
s3Bucket.Put. In https://github.com/aws/aws-sdk-go-v2/discussions/2960,
the AWS SDK changed to automatically default to this algorithm for uploads.
but it does not seem to default correctly on multi-part completions
when using the Uploader.Uploader API. This caused errors like

  api error InvalidPart: One or more of the specified parts could not be
  found. The part may not have been uploaded, or the specified entity tag
  may not match the part's entity tag.

when pail was updated in Evergreen. We didn't catch this in pail's tests
because it appears that the behavior was only changed on upgrading the
aws config package in addition to the S3 SDK, at least as far as I could
tell. To fix this, I have updated all of the AWS SDK libraries here to
match what is in Evergreen right now.