DEVPROD-17133: correctly close gzip buffer on put
This commit fixes a bug when Compress: true was passed to pail where we
forgot to close the gzip writer. This bug was introduced in 10686cd7fb4e407b49a2099d769331646b6fa372.
The tests did not catch this mistake because we were only testing the write
path through .Writer instead of .Put. This commit also adds a test for Put
which uses compression. Without the corresponding change to s3_bucket.go,
this test case will fail as expected.